Changed file organisation
